Encapsulate IAMSigner in request builder
