+++ src/main/scala/com/gu/crossword/crosswords/RequestBuilder.scala
@@ -1,5 +1,7 @@
 package com.gu.crossword.crosswords
 
+import com.amazonaws.auth.profile.ProfileCredentialsProvider
+import com.amazonaws.auth.{AWSCredentialsProvider, AWSCredentialsProviderChain, STSAssumeRoleSessionCredentialsProvider}
 import com.gu.contentapi.client.IAMSigner
 import okhttp3.{Headers, Request}
 
@@ -7,13 +9,25 @@ import java.net.URI
 import scala.jdk.CollectionConverters._
 
 trait RequestBuilder {
-  val builder: Request.Builder
-
   def buildRequest(url: String, withAuth: Boolean): Request
 }
 
-class RequestBuilderWithSigner(signer: IAMSigner) extends RequestBuilder {
-  val builder = new Request.Builder()
+class RequestBuilderWithSigner(capiPreviewRole: String, awsRegion: String) extends RequestBuilder {
+  private val builder = new Request.Builder()
+
+  private val signer = {
+    val capiPreviewCredentials: AWSCredentialsProvider = {
+      new AWSCredentialsProviderChain(
+        new ProfileCredentialsProvider("capi"),
+        new STSAssumeRoleSessionCredentialsProvider.Builder(
+          capiPreviewRole,
+          "capi"
+        ).build()
+      )
+    }
+
+    new IAMSigner(capiPreviewCredentials, awsRegion)
+  }
 
   def buildRequest(url: String, withAuth: Boolean) = {
     val req = builder
@@ -31,7 +45,7 @@ class RequestBuilderWithSigner(signer: IAMSigner) extends RequestBuilder {
 }
 
 class BasicRequestBuilder extends RequestBuilder {
-  val builder = new Request.Builder()
+  private val builder = new Request.Builder()
 
   def buildRequest(url: String, withAuth: Boolean) = builder.url(url).build
 }
